REFACTOR-PASSPORT/Probando nueva configuración
