+++ config/passport-setup.js
@@ -16,8 +16,6 @@ passport.serializeUser((user, done) => {
   console.log(`User id = ${user.id}`);
   console.log(`User displayName: ${user.displayName}`);
   //le paso el id que crea la DB, y NO la id de google.
-  
-  
   done(null, user.id);
   
 });
